% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown.R
\name{countdown}
\alias{countdown}
\title{countdown}
\usage{
countdown(start, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{start}{countdown start date}

\item{width}{widget width}

\item{height}{widget height}

\item{elementId}{widget id}
}
\description{
Countdown widget
}
